\name{leverageCycles-package}
\alias{leverageCycles-package}
\alias{leverageCycles}
\docType{package}
\title{
\packageTitle{leverageCycles}
}
\description{
\packageDescription{leverageCycles}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{leverageCycles}
\packageIndices{leverageCycles}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{leverageCycles}

Maintainer: \packageMaintainer{leverageCycles}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
